Relocate getCatalogingLanguage() to utils.js. Fix references to it.
