+++ src/reducers/utils.js
@@ -275,3 +275,15 @@ export function stringToRegex(string) { // easier to remember
   return new RegExp(string, 'u');
 }
 */
+
+export function getCatalogingLanguage(record) {
+  const [field040] = record.get(/^040$/u);
+  if (!field040) {
+    return null;
+  }
+  const [b] = field040.subfields.filter(sf => sf.code === 'b');
+  if (!b) {
+    return null;
+  }
+  return b.value;
+}
